api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.5.2
huggingface-hub==0.19.4
sentence-transformers==2.2.2  # Fixed typo in the package name
python-multipart==0.0.6
requests==2.31.0
pytz==2023.3
newsapi-python==0.2.7
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.25  # Added for Supabase integration
asyncpg==0.29.0  # Added for async PostgreSQL operations
pgvector==0.2.4  # Added for vector similarity search
supabase==1.0.4  # Added for Supabase integration
aiohttp==3.9.1  # Added for async HTTP requests (Upstash REST API)
pandas==2.1.3
scikit-learn==1.3.2
torchaudio>=2.2.0  # Updated to be compatible with arm64
torchvision>=0.17.0  # Updated to be compatible with arm64
torch>=2.2.0  # Updated to be compatible with arm64
bs4==0.0.1
